[Android] Network 관련 코드 분리 및 레포지토리 이름 변경 (#45)
* refact: 네트워크 관련 코드 분리

History뿐 아니라 Todo에서도 네트워크를 사용해야 함으로 공통으로 쓸 수 있도록 분리

* refact: HistoryRepository를 공용으로 사용할 수 있도록 위치 및 파일명 변경
