--- Android/app/src/main/java/com/example/todo_list/history/data/HistoryRepository.kt
@@ -1,13 +1,0 @@
-package com.example.todo_list.history.data
-
-import com.example.todo_list.Repository
-import com.example.todo_list.history.HistoryReceive
-import retrofit2.Response
-
-class HistoryRepository : Repository {
-    private val retrofit = HistoryReceive.service
-
-    override suspend fun getHistories(): Response<List<HistoryCard>> {
-        return retrofit.getHistories("histories")
-    }
-}
