+++ Android/app/src/main/java/com/example/todo_list/network/TodoService.kt
@@ -1,0 +1,14 @@
+package com.example.todo_list.network
+
+import com.example.todo_list.history.data.HistoryCard
+import retrofit2.Response
+import retrofit2.http.GET
+import retrofit2.http.Path
+
+interface TodoService {
+
+    @GET("api/{histories}")
+    suspend fun getHistories(
+        @Path("histories") variable: String
+    ): Response<List<HistoryCard>>
+}
